+++ src/Omega/Container/Exceptions/AbstractContainerException.php
@@ -1,0 +1,39 @@
+<?php
+
+/**
+ * Part of Omega - Container Package.
+ *
+ * @link      https://omega-mvc.github.io
+ * @author    Adriano Giovannini <agisoftt@gmail.com>
+ * @copyright Copyright (c) 2026 Adriano Giovannini (https://omega-mvc.github.io)
+ * @license   https://www.gnu.org/licenses/gpl-3.0-standalone.html     GPL V3.0+
+ * @version   1.0.0
+ */
+
+declare(strict_types=1);
+
+namespace Omega\Container\Exceptions;
+
+use RuntimeException;
+
+/**
+ * Base exception for all dependency injection container errors.
+ *
+ * All container-specific exceptions extend this class, providing a common
+ * exception type while remaining compatible with
+ *
+ * @category   Omega
+ * @package    Container
+ * @subpackage Exceptions
+ * @link       https://omega-mvc.github.io
+ * @author     Adriano Giovannini <agisoftt@gmail.com>
+ * @copyright  Copyright (c) 2026 Adriano Giovannini (https://omega-mvc.github.io)
+ * @license    https://www.gnu.org/licenses/gpl-3.0-standalone.html     GPL V3.0+
+ * @version    1.0.0
+ *
+ * {@see ContainerExceptionInterface}.
+ */
+class AbstractContainerException extends RuntimeException implements ContainerExceptionInterface
+{
+}
+