+++ src/Omega/Container/Exceptions/ClassNotFoundException.php
@@ -1,0 +1,46 @@
+<?php
+
+/**
+ * Part of Omega - Container Package.
+ *
+ * @link      https://omega-mvc.github.io
+ * @author    Adriano Giovannini <agisoftt@gmail.com>
+ * @copyright Copyright (c) 2026 Adriano Giovannini (https://omega-mvc.github.io)
+ * @license   https://www.gnu.org/licenses/gpl-3.0-standalone.html     GPL V3.0+
+ * @version   1.0.0
+ */
+
+declare(strict_types=1);
+
+namespace Omega\Container\Exceptions;
+
+use function sprintf;
+
+/**
+ * Thrown when a requested class cannot be found or loaded.
+ *
+ * @category   Omega
+ * @package    Container
+ * @subpackage Exceptions
+ * @link       https://omega-mvc.github.io
+ * @author     Adriano Giovannini <agisoftt@gmail.com>
+ * @copyright  Copyright (c) 2026 Adriano Giovannini (https://omega-mvc.github.io)
+ * @license    https://www.gnu.org/licenses/gpl-3.0-standalone.html     GPL V3.0+
+ * @version    1.0.0
+ */
+class ClassNotFoundException extends AbstractContainerException
+{
+    /**
+     * Create a new exception instance.
+     *
+     * @param string $className Fully-qualified class name that could not be loaded.
+     * @return void
+     */
+    public function __construct(string $className)
+    {
+        parent::__construct(
+            sprintf('Unable to load class "%s".', $className)
+        );
+    }
+}
+