+++ src/Omega/Container/Exceptions/ContainerExceptionInterface.php
@@ -1,0 +1,38 @@
+<?php
+
+/**
+ * Part of Omega - Container Package.
+ *
+ * @link      https://omega-mvc.github.io
+ * @author    Adriano Giovannini <agisoftt@gmail.com>
+ * @copyright Copyright (c) 2026 Adriano Giovannini (https://omega-mvc.github.io)
+ * @license   https://www.gnu.org/licenses/gpl-3.0-standalone.html     GPL V3.0+
+ * @version   1.0.0
+ */
+
+declare(strict_types=1);
+
+namespace Omega\Container\Exceptions;
+
+use Throwable;
+
+/**
+ * Base interface for all container-related exceptions.
+ *
+ * Every exception thrown by the dependency injection container implements
+ * this interface, allowing consumers to catch container-specific failures
+ * independently of generic runtime exceptions.
+ *
+ * @category   Omega
+ * @package    Container
+ * @subpackage Exceptions
+ * @link       https://omega-mvc.github.io
+ * @author     Adriano Giovannini <agisoftt@gmail.com>
+ * @copyright  Copyright (c) 2026 Adriano Giovannini (https://omega-mvc.github.io)
+ * @license    https://www.gnu.org/licenses/gpl-3.0-standalone.html     GPL V3.0+
+ * @version    1.0.0
+ */
+interface ContainerExceptionInterface extends Throwable
+{
+}
+