+++ src/Omega/Container/Exceptions/RecursiveDependencyException.php
@@ -1,0 +1,49 @@
+<?php
+
+/**
+ * Part of Omega - Container Package.
+ *
+ * @link      https://omega-mvc.github.io
+ * @author    Adriano Giovannini <agisoftt@gmail.com>
+ * @copyright Copyright (c) 2026 Adriano Giovannini (https://omega-mvc.github.io)
+ * @license   https://www.gnu.org/licenses/gpl-3.0-standalone.html     GPL V3.0+
+ * @version   1.0.0
+ */
+
+declare(strict_types=1);
+
+namespace Omega\Container\Exceptions;
+
+use function sprintf;
+
+/**
+ * Thrown when the container detects a circular dependency.
+ *
+ * @category   Omega
+ * @package    Container
+ * @subpackage Exceptions
+ * @link       https://omega-mvc.github.io
+ * @author     Adriano Giovannini <agisoftt@gmail.com>
+ * @copyright  Copyright (c) 2026 Adriano Giovannini (https://omega-mvc.github.io)
+ * @license    https://www.gnu.org/licenses/gpl-3.0-standalone.html     GPL V3.0+
+ * @version    1.0.0
+ */
+class RecursiveDependencyException extends AbstractContainerException
+{
+    /**
+     * Create a new exception instance.
+     *
+     * @param string $identifier Identifier involved in the circular dependency.
+     * @return void
+     */
+    public function __construct(string $identifier)
+    {
+        parent::__construct(
+            sprintf(
+                'Circular dependency detected while resolving "%s".',
+                $identifier
+            )
+        );
+    }
+}
+