+++ CHANGELOG.md
@@ -3,3 +3,22 @@
 ### Initial Version
 
 - Welcome to Omega Framework For WordPress
+
+## [1.0.0] - 2026-07-19
+
+### Added
+
+ - Adding container exception.
+ - Adding container interface.
+
+### Changed
+
+- Refactoring the container class.
+
+### Fixed
+
+- None.
+
+### Removed
+
+- None.